+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - [#229] Allow to application manage signing key and algorithm
 - [#230] Add dynamic client registration
+- [#233] fix: handle DoubleRenderError in library instead of requiring consumer workaround
 
 ## v1.8.11 (2025-02-10)