+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [#230] Add dynamic client registration
 - [#233] fix: handle DoubleRenderError in library instead of requiring consumer workaround
 - [#232] Implements customizable OpenID request class
+- [#236] Derive token_endpoint_auth_methods_supported from Doorkeeper's client_credentials config
 
 ## v1.8.11 (2025-02-10)