+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [#233] fix: handle DoubleRenderError in library instead of requiring consumer workaround
 - [#232] Implements customizable OpenID request class
 - [#236] Derive token_endpoint_auth_methods_supported from Doorkeeper's client_credentials config
+- [#225] Allow configuration of id_token expiration using a block.
 
 ## v1.8.11 (2025-02-10)